airflow_requirements.txt
# Airflow core
apache-airflow==2.8.0
apache-airflow-providers-amazon==8.15.0

# ML dependencies from your existing pipeline
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=9.5.0
tqdm>=4.65.0

# AWS and data handling
boto3>=1.26.0
botocore>=1.29.0
python-dotenv>=1.0.0

# MLflow for experiment tracking
mlflow>=2.7.0

# Additional utilities
scipy>=1.10.0

# For sentiment analysis (from your preprocessing)
# Note: Make sure this model is available in your environment
tabularisai-multilingual-sentiment-analysis 